adding in the decorator package
it is a dependency of `invoke`, which is a dependency of `fabric`. in
the 2.1.0 version `invoke`, the `decorator` package was removed. the
reason is not clear, but the `invoke` package maintainer plans to add it
back in for one of the 3.x versions.

more infor at https://github.com/pyinvoke/invoke/issues/935